[lib] Move platform configuration to new platform/config package
This moves the platform configuration functions to a new package
platform/config. The goal of this change is to make is so that changes
to unrelated parts of the platform package don't rebuild the entire
world.
