+++ .bldr.toml
@@ -29,8 +29,8 @@ paths = [
   "components/pg-sidecar-service/*",
   "lib/grpc/*",
   "lib/io/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -113,8 +113,8 @@ paths = [
   "lib/io/*",
   "lib/logger/*",
   "lib/oidc/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -187,8 +187,8 @@ paths = [
   "lib/grpc/*",
   "lib/io/*",
   "lib/logger/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -500,8 +500,8 @@ paths = [
   "lib/io/*",
   "lib/logger/*",
   "lib/pgutils/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -999,6 +999,7 @@ paths = [
   "lib/io/*",
   "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/pg/*",
   "lib/platform/sys/*",
   "lib/proc/*",
@@ -1140,8 +1141,8 @@ paths = [
   "lib/io/*",
   "lib/logger/*",
   "lib/pgutils/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -1230,8 +1231,8 @@ paths = [
   "lib/io/*",
   "lib/logger/*",
   "lib/pgutils/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -1326,8 +1327,8 @@ paths = [
   "lib/event/*",
   "lib/grpc/*",
   "lib/io/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -1393,8 +1394,8 @@ paths = [
   "lib/io/*",
   "lib/license/*",
   "lib/logger/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -1508,8 +1509,8 @@ paths = [
   "lib/io/*",
   "lib/logger/*",
   "lib/pgutils/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -1574,8 +1575,8 @@ paths = [
   "lib/io/*",
   "lib/logger/*",
   "lib/oidc/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -1645,8 +1646,8 @@ paths = [
   "lib/grpc/*",
   "lib/io/*",
   "lib/logger/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
@@ -1922,8 +1923,8 @@ paths = [
   "lib/cereal/*",
   "lib/grpc/*",
   "lib/io/*",
-  "lib/platform/*",
   "lib/platform/command/*",
+  "lib/platform/config/*",
   "lib/platform/sys/*",
   "lib/proc/*",
   "lib/proxy/*",
