+++ components/authz-service/cmd/authz-service/main.go
@@ -16,7 +16,7 @@ import (
 	"github.com/chef/automate/components/authz-service/storage/postgres/migration"
 	"github.com/chef/automate/lib/grpc/secureconn"
 	"github.com/chef/automate/lib/logger"
-	"github.com/chef/automate/lib/platform"
+	platform_config "github.com/chef/automate/lib/platform/config"
 	"github.com/chef/automate/lib/tls/certs"
 	"github.com/chef/automate/lib/tracing"
 )
@@ -64,7 +64,7 @@ Please pass a config file as the only argument to this command.`))
 
 	if cfg.PGURL == "" {
 		var err error
-		cfg.PGURL, err = platform.PGURIFromEnvironment(cfg.Database)
+		cfg.PGURL, err = platform_config.PGURIFromEnvironment(cfg.Database)
 		if err != nil {
 			fail(errors.Wrap(err, "Failed to get pg uri"))
 		}
